Dye Salvaging, Silveress for history

+++ GW2Tradz/GW2Tradz/Viewmodels/TradingAction.cs
@@ -11,7 +11,7 @@ namespace GW2Tradz.Viewmodels
         public string Description{get; set; }
         public Item Item { get; set; }
         public int Profit { get; set; }
-        public float ProfitPercentage { get; set; }
+        public double ProfitPercentage { get; set; }
         public int Amount { get; set; }
     }
 }
